Update chainparams.cpp
Prune height increased from 100,000 to 500,000.
This is because most blocks are empty. 
As of height 143516 after 6 months, the blockchain size is 93.6 MB.
Assuming a doubling every 100,000 blocks, the blockchain will still be well under 2 GB.

The space savings from disabling pruning is not worth resynchronising in the case the full chain is needed.